Modified entity classes, added unit testing

+++ src/main/java/com/example/main/topology/component/nmos/M1.java
@@ -11,7 +11,7 @@ import lombok.*;
 
 public class M1 {
     @JsonProperty("default")
-    private int defaultValue;
+    private float defaultValue;
     private int min;
     private int max;
 }
